[Feature #1333] Create package for validation-related code. Move ValidationResult into it.
