+++ src/main/java/cz/cvut/kbss/termit/persistence/dao/changetracking/ChangeTrackingHelperDao.java
@@ -24,7 +24,7 @@ public class ChangeTrackingHelperDao {
      * @param update Current state of the asset to find
      * @return Stored state of the searched asset
      */
-    public <T extends Asset> T findStored(T update) {
+    public <T extends Asset<?>> T findStored(T update) {
         Objects.requireNonNull(update);
         final T result = (T) em.find(update.getClass(), update.getUri());
         if (result == null) {
