+++ src/main/java/cz/cvut/kbss/termit/validation/ValidationResult.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
-package cz.cvut.kbss.termit.util;
+package cz.cvut.kbss.termit.validation;
 
 import javax.validation.ConstraintViolation;
 import java.io.Serializable;
